feat: show exporter URL
