+++ src/app/app.go
@@ -4,6 +4,7 @@ import (
 	"flag"
 	"fmt"
 	"os"
+	"qbit-exp/internal"
 	"qbit-exp/logger"
 	"strconv"
 	"strings"
@@ -26,6 +27,7 @@ type ExporterSettings struct {
 	LogLevel            string
 	ExperimentalFeature ExperimentalFeatures
 	Feature             Features
+	URL                 string
 }
 
 type QBittorrentSettings struct {
@@ -71,6 +73,7 @@ func LoadEnv() {
 	enableTracker := getEnv(defaultDisableTracker)
 	enableHighCardinality := getEnv(defaultHighCardinality)
 	labelWithHash := getEnv(defaultLabelWithHash)
+	exporterUrl := getEnv(defaultExporterURL)
 
 	exporterPort, errExporterPort := strconv.Atoi(exporterPortEnv)
 	if errExporterPort != nil {
@@ -86,6 +89,10 @@ func LoadEnv() {
 	}
 	if timeoutDuration < 0 {
 		panic(fmt.Sprintf("%d must be > 0", timeoutDuration))
+	}
+
+	if exporterUrl != "" && !internal.IsValidURL(exporterUrl) {
+		panic(fmt.Sprintf("%s is not a valid URL", exporterUrl))
 	}
 
 	QBittorrent = QBittorrentSettings{
@@ -106,6 +113,7 @@ func LoadEnv() {
 		},
 		LogLevel: loglevel,
 		Port:     exporterPort,
+		URL:      exporterUrl,
 	}
 
 }
