+++ src/internal/internal.go
@@ -1,0 +1,8 @@
+package internal
+
+import "net/url"
+
+func IsValidURL(input string) bool {
+	u, err := url.Parse(input)
+	return err == nil && u.Scheme != "" && u.Host != ""
+}
